Revert "refactor: utils to multi-files (#86)"
This reverts commit 6f9d83c4f35e574e195a027783a6cf4ff16fc632.
