+++ app/(routes)/(main)/blog/page.tsx
@@ -2,10 +2,10 @@ import { Badge } from '@/components/ui/badge';
 import { Separator } from '@/components/ui/separator';
 import { db } from '@/db';
 import { posts, PostSelect } from '@/db/schema';
+import { formatDate } from '@/lib/utils';
 import Link from 'next/link';
 import { Metadata } from 'next';
 import { desc } from 'drizzle-orm';
-import { formatDate } from '@/utils/date';
 
 export const metadata: Metadata = {
   title: 'Blog | Moonlitgrace',
