--- utils/date.ts
@@ -1,9 +1,0 @@
-export function formatDate(dateStr: Date) {
-  const date = new Date(dateStr).toLocaleDateString('en-EN', {
-    day: 'numeric',
-    month: 'short',
-    year: 'numeric',
-  });
-  return date;
-}
-
