--- utils/file.ts
@@ -1,4 +1,0 @@
-export function validateFile(file: File) {
-  return file instanceof File && file.name !== '' && file.size > 0;
-}
-
