--- utils/string.ts
@@ -1,4 +1,0 @@
-export function truncate(str: string, n: number) {
-  return str.length > n ? str.slice(0, n - 3) + '...' : str;
-}
-
